requirements_inference.txt
hydra-core==1.3.2
numpy==1.26.3
omegaconf==2.3.0
pandas==2.1.4
pytorch_lightning==2.1.2
scikit_learn==1.3.2
torch==2.1.2  
transformers==4.36.2
tensorboard==2.15.1
wandb==0.16.2
streamlit>=1.30.0
flask==3.0.0